test(checkAddresses): split into several files
